chore: moves and renames the component
